# src/api_docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Blog",
        "description": "Personal Blog API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "basePath": "/",
...
    "schemes": ["https", "http"],
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "paths":{
...
        "/posts": {
            "get": {
                "tags": ["Post"],
                "summary": "Get Posts of current user",
                "description": "List all the user posts",
                "produces": ["application/json"],
                "security": [
                    { 
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "$ref": "#/parameters/Title"
                    },
                    {
                        "$ref": "#/parameters/Size"
                    },
                    {
                        "$ref": "#/parameters/Page"
                    },
                    {
                        "$ref": "#/parameters/Order"
                    },
                    {
                        "$ref": "#/parameters/OrderBy"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfull operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Post"
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["Post"],
                "summary": "Add Post",
                "description": "",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "security": [
                    { 
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Post details",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Post"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Invalid Body Parameters"
                    }
                }
            }
        },
        "/posts/{postId}": {
            "put": {
                "tags": ["Post"],
                "summary": "Update Post",
                "description": "",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "security": [
                    { 
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Post details",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Post"
                        }
                    },
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "Identifier of the Post",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid Body Parameters"
                    },
                    "404": {
                        "description": "Identifier not found."
                    }
                }
            },
            "delete": {
                "tags": ["Post"],
                "summary": "Delete Post",
                "description": "",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "security": [
                    { 
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "Identifier of the Post",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Identifier not found."
                    }
                }
            }
        },
...
    },
    "parameters": {
...
        "Title": {
            "name": "title",
            "in": "query",
            "description": "filter posts by title",
            "type": "string"
        },
...
        "Order": {
            "name": "order",
            "in": "query",
            "description": "ASC or DESC",
            "type": "string",
            "enum": [
                "ASC",
                "DESC"
            ]
        },
        "OrderBy": {
            "name": "orderBy",
            "in": "query",
            "description": "Order values that need to be considered for filter",
            "type": "string"
        },
        "Size": {
            "name": "size",
            "in": "query",
            "description": "Limit of items for a page",
            "type": "number"
        },
        "Page": {
            "name": "page",
            "in": "query",
            "description": "Current number of the page",
            "type": "number"
        }
    },
    "definitions": {
...
        "Post": {
            "type": "object",
            "required": ["title", "content", "category"],
            "properties": {
                "title": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "category": {
                    "type": "object",
                    "properties": {
                        "categoryId": {
                            "type": "string",
                            "required": true
                        },
                        "name": {
                            "type": "string"
                        }
                    }
                },
                "author": {
                    "type": "object",
                    "properties": {
                        "userId": {
                            "type": "string"
                        },
                        "name": {
                            "type": "string"
                        },
                        "email": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
}